Bookmark saving verification + deleteBookmark moved
Placed a status verifier when the bookmarks is saved and for when it's a success and for when there is an error. Error has not been checked yet since I dont know how to cause an error whilst saving a bookmark. Function for deleting bookmark was moved fromt he popupHTML script to the bookmarkScript for better modularity and clarity